+++ apps/settings.h
@@ -41,7 +41,7 @@
 
 struct opt_items {
     unsigned const char* string;
-    long voice_id;
+    int32_t voice_id;
 };
 
 /** Setting values defines **/
@@ -254,7 +254,7 @@ bool set_int_ex(const unsigned char* string, const char* unit, int voice_unit,
              int* variable,
              void (*function)(int), int step, int min, int max, 
              void (*formatter)(char*, size_t, int, const char*),
-             long (*get_talk_id)(int));
+             int32_t (*get_talk_id)(int));
 
 /* the following are either not in setting.c or shouldnt be */
 bool set_time_screen(const char* string, struct tm *tm);
